Add in songs routing
