+++ routes/playlist.js
@@ -1,6 +1,5 @@
 const express = require("express");
 const { AppError } = require("../src/errors/AppError");
-const app = require("../app");
 const router = express.Router();
 
 let playlists = [
@@ -53,36 +52,6 @@ router.get("/:id", (req, res) => {
   const { id } = req.params;
   const playlist = playlistGetByID(id);
   res.status(200).json(playlist);
-});
-
-router.get("/:id/songs", (req, res) => {
-  const { id } = req.params;
-  const playlist = playlistGetByID(id);
-  const songList = songListObjectGetByID(playlist.id);
-  res.status(200).json([...songList.songArray]);
-});
-
-router.patch("/:id/songs/:songID", (req, res) => {
-  const { id, songID } = req.params;
-  const body = req.body;
-  const playlist = playlistGetByID(id);
-  const songList = songListObjectGetByID(playlist.id);
-  const songListArray = songList.songArray;
-
-  let updatedSong;
-
-  const updatedSongListArray = songListArray.map((song) => {
-    if (song.id === songID) {
-      updatedSong = { ...song, ...body };
-      return updatedSong;
-    }
-    return song;
-  });
-
-  if (!updatedSong) throw new AppError(404, "Song ID not found");
-  songList.songArray = updatedSongListArray;
-
-  res.status(200).json(updatedSong);
 });
 
 module.exports = router;
